move _extra tags into _tags file
