+++ styleguide/templates/styleguide/styleguide.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% load styleguide_extras %}
+{% load styleguide_tags %}
 
 {% block content %}
 <link rel=stylesheet href="/static/styleguide/highlightjs/styles/github.css" />
