+++ styleguide/templatetags/styleguide_tags.py
@@ -1,7 +1,11 @@
+import re
+
 from textwrap import dedent
 
 from django import template
 from django.template import defaultfilters
+
+from styleguide import utils
 
 
 register = template.Library()
@@ -41,3 +45,16 @@ class ExampleNode(template.Node):
         output.append("</div></div>")
 
         return ''.join(output)
+
+
+@register.assignment_tag
+def get_styleguide_templates():
+    """Return tuples of (display name, slug) for found styleguide templates"""
+    templates = []
+    for template_ in utils.get_styleguide_templates():
+        match = re.search(r'^styleguide-([\w-]+)\.html$', template_)
+        if match:
+            slug = match.group(1)
+            name = defaultfilters.title(slug.replace('-', ' '))
+            templates.append((name, slug))
+    return templates
